Hunting for warnings in "Warning hunting" :p
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@5681 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/MatchObjects.cpp
@@ -282,22 +282,22 @@ void CMatchList::AddHits(CQueryHit* pHit, CQuerySearch* pFilter, BOOL bRequire)
 			
 			if ( Stats.nHadCount )
 			{
-				ASSERT( m_nItems >= Stats.nHadCount );
+				ASSERT( m_nItems >= (UINT)Stats.nHadCount );
 				m_nItems -= Stats.nHadCount;
 				ASSERT( m_nFilteredFiles );
 				m_nFilteredFiles --;
-				ASSERT( m_nFilteredHits >= Stats.nHadFiltered );
+				ASSERT( m_nFilteredHits >= (UINT)Stats.nHadFiltered );
 				m_nFilteredHits -= Stats.nHadFiltered;
 
 				switch ( nProtocol )
 				{
 				case PROTOCOL_G1:
 				case PROTOCOL_G2:
-					ASSERT( m_nGnutellaHits >= Stats.nHadFiltered );
+					ASSERT( m_nGnutellaHits >= (UINT)Stats.nHadFiltered );
 					m_nGnutellaHits -= Stats.nHadFiltered;
 					break;
 				case PROTOCOL_ED2K:
-					ASSERT( m_nED2KHits >= Stats.nHadFiltered );
+					ASSERT( m_nED2KHits >= (UINT)Stats.nHadFiltered );
 					m_nED2KHits -= Stats.nHadFiltered;
 					break;
 				case PROTOCOL_BT:
